[SettingsLib] Dynamically update StatusBannerPreference icon tint
Prior to this change, the icon tint was only set during initialization
based on the icon or icon level. This allows the icon tint to be
dynamically updated based on a respective icon level change. Setting
the icon level whenthere is no icon defined is a no-op.

Bug: 400999418
Test: Manual
Flag: EXEMPT bugfix
Change-Id: I37d4cf890ecf9152013ac5c53cb60bc55fb26643