Prevent SettingsPreferenceGroupAdapter from changing StatusBannerPreference background
When the expressive theme is enabled, StatusBannerPreference loses its
background color. This is because SettingsPreferenceGroupAdapter will
override the preference's inner MaterialCardView background.

Avoid this issue by wrapping the MaterialCardView in
StatusBannerPreference with a FrameLayout which will prevent
SettingsPreferenceGroupAdapter from modifying its background as part of
the expressive theme update.

This is syncs the changes from cl/739045958. Screenshot:
http://screen/3USsVw2V8oSFLFS.png

Bug: 400986976
Test: Manually verify UI. See screenshot.
Flag: EXEMPT bugfix
Change-Id: I094ddeca7435cc9a5c32b2ed51e537a08d5fbe8c