Only update remote target, if visibility is different from input target
When syncing the visibility of the imeInputTarget to the (RemoteInsets)ControlTarget, without having an actual change, the upcoming call to updateClientVisibility will be ignored and we don't invoke the IMMS listener.
Therefore, we should only sync the visibility when it's different between imeInputTarget and controlTarget - otherwise we can directly invoke the listener.

Fix: 402547338
Fix: 405296840
Flag: android.view.inputmethod.refactor_insets_controller
Test: atest ImeInsetsSourceProviderTest#testUpdateControlForTarget_remoteInsetsControlTargetUnchanged
Test: manual: open view where controlTarget=RemoteInsetsControlTarget; Open another fullscreen app and request IME; close new window to return to RICT
Change-Id: I96ffe688ae5eb32e325f3f7c9c62b392f7287100