Update dialog visibility in SysUIState per display
Many classes in launcher hide UI element when there is a dialog visible in SysUI.
This propagates the "dialog visible" bit in the correct SysUIState associated with a display, resulting in launcher hiding the correct elements.

Before this, for example, opening a dialog in the external display was causing the taskbar to be hidden in the default display.

Note the context will have a different display id only when
shade_window_goes_around is enabled, so this is essentially guarded by
that flag.

Bug: 362719719
Bug: 405959656
Test: SystemUIDialogTest, SysUIStatePerDisplayInteractorTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I114d568f1d69f8f74820abae94706a76253f6dd5