Fix system dream component race condition
There's a delay between when the doze dream starts and when it notifies DreamManager that its dozing. If setSystemDreamComponent is called during this delay, then we may enter the lowlight dream from the doze dream.

To avoid this, we check if the current dream can doze at all, which is only true for the doze dream. This gives us the signal that its a doze dream sooner than the actual isDozing check.

Bug: 404968947
Test: manually by forcing lowlight using adb
Flag: EXEMPT bugfix
Change-Id: I77fc178b6a4e014be051966188f593ae0ecea93e