Make leash cleanup on-demand instead of default.
For some reason, Launcher still has performance regressions if we
always do this. I'm not sure why that is, but I don't have the
bandwidth to investigate right now. Since the change was to avoid a
leak when using the Animation library in the first place, we add an
explicit parameter that is false by default so we can limit the
behavior change to the library.

Bug: 403833923
Flag: EXEMPT bugfix
Test: manual
Change-Id: Iec8e107ff5fa05d3ab7f353218985b6eba26d817