fix(EDT): Timing of ForceInvert state usage
Problem: Both ViewRootImpl and UIModeManager register for the same
settings key related to ForceInvert state changes. The unpredictable
order in which their respective change callbacks are executed can
lead to the "force invert" functionality not working as expected in
certain scenarios.

Solution: To ensure consistent force inversion behavior, we should
rely solely on the UIModeManager's callback to handle changes. The
registration for this settings key in ViewRootImpl will be removed.
This guarantees that the theme change logic within UIModeManager is
always executed, maintaining the intended order of operations and
resolving the inconsistent inversion issue.

NO_IFTTT=Remove the tag

Bug: 399642883
Flag: android.view.accessibility.force_invert_color
Test: atest ViewRootImplTest
Test: Manual Fitbit/A11y FAB test for patch
Test: Manual testing that no prior Launcher memory leak
Test: ABTD for no cause SysUI blocking call regression
Change-Id: I6e7e03b653b81216a77997c062b873dc4243aa87