Fix for missing widgets
This works around around a Jetpack Compose bug where if we add a child
to the element inside an AndroidView composable before the element has
been laid out, it will fail to render. We therefore delay processing the
RemoteViews for the widget until after the parent has been laid out.

Test: manually on device by verifying widgets properly show up
Flag: EXEMPT bugfix
Bug: 387938328

Change-Id: I3b094086dd133981c37fa9709547334091fcb580