[Dual Shade] Fix blur and depth effects in scene container.
`NotificationShadeDepthController` is the class responsible for
controlling the blur and depth effects for the wallpaper. While it is
part of the legacy code (pre-scene container), it is still active in
scene container too, yet its state was partially left stale in the
latter.

BONUS: Remove `legacyPanelExpansion` from `PanelExpansionInteractor` and
`NotificationPanelViewController`, as both are no longer used after this
CL.

Summary of the depth controller state that is being changed in this CL:

- `legacyShadeExpansion` was marking the shade as fully expanded when
  transitioning to the lockscreen, which meant we were not unblurring
  during that transition. It is now replaced with `shadeExpansion`.

- `qsPanelExpansion` was only being set from the legacy
  `QuickSettingsControllerImpl`, which isn't active in scene container.
  It is now set based on `shadeInteractor`.

- `transitionToFullShadeProgress` was only being set from the legacy
  `LockscreenShadeTransitionController`, which isn't active in scene
  container. It is now set based on `shadeInteractor`.

Bug: 388470180
Bug: 370556579
Test: Updated unit tests.
Test: Manually by opening the notifications shade and quick settings
 shade over both lock screen and launcher, and verifying correct blur
 behavior during the transitions, incl. switching between the shades.
Flag: com.android.systemui.scene_container
Change-Id: Ie66ea3e358441c14d9ef22373ac4e5a976d15638