DisplayTopology: remove density param of getGraph
This density map is no longer needed since TreeNode objects have a
logical density and logical px size fields.

This includes a cleanup of proximate code in DisplayTopology which I had
wanted to do for a while. The getInfo method now returns a single List
with a new record type, rather than build a HashMap for each desired
piece of information.

I estimate the List is around as memory-heavy as a single HashMap, as a
HashMap requires an allocation for each Entry and the List requires an
allocation for each NodeDerivedInfo. In this CL, the the RectF bounds
objects are constructed on-demand.

More importantly, the changes were done for the sake of maintainability
and there should be no effect on performance.

Flag: EXEMPT refactor
Bug: b/407028459
Test: DisplayTopologyTest DisplayManagerServiceTest DisplayTopologyCoordinatorTest
Change-Id: I2aeab51d6f43254a438efb4afdba1107380d5533