Change how we set NotificationBackgroundView colors for row transparency
The current implementation uses color filter with the mode SRC. This replaces all the original pixels in both color and alpha. It causes a jump cut as the view dims out from collapsing of the shade, since the dimming doesn't do anything with the color filter.

Since the base layer of the background is a gradient drawable, directly manipulating the base layer drawable color seems to work. It's transparent and dims out when the shade is collapsed

Fixes: 399849326
Test: manual
Flag: com.android.systemui.notification_row_transparency
Change-Id: I5f75f5bdc68ec2cef4e932d8f03be76519b60dec