Fix weather complication not showing after wipe flash
After a wipe install, the weather complication may not appear on the
dream overlay. It is due to the addListenerForWeatherPlugin call
occurring on system start and if the lockscreen smartspace precondition
isn't met (true before the device is set up), then the session to dream
surface is not created.

Fix is to create the session within the smartspace precondition changed
listener if one doesn't already exist.

Fixes: b/382757398
Flag: EXEMPT bugfix
Test: wipe install, finish SUW and verify weather showing on dream
Test: atest DreamSmartspaceControllerTest
Change-Id: I8ee86c565ec29566a3393e30e6d6ee95032e982e