Create KCM copy when manipulating contents per language
Earlier we were using Virtual KCM and applying an overlay
KCM on it which was updating the keys map (from the
caller thread) which can corrupt the data.

We should not touch Virtual KCM file which can be
accessed from multiple processes, just create a copy
and manipulate the copy instead.

Test: manual
Bug: 407036640
Flag: EXEMPT bugfix
Change-Id: I50f1d142ea7a2eac3ed613a2d9cae3acd8bba354