[SB][ComposeIcons] Create Composable System Status Icons
This change sets the foundation for system status icons in compose
by doing the following:

Adds a composable to contain system status icons in the statusbar.
Currently, the container is showing up in the middle of the screen
because it is filling the width of status_bar_end_side_container and
the content is aligned to the start.

Adds a SystemStatusIconsViewModel which will consolidate individual
SystemStatusIconViewModels to emit a list of icons to be displayed
in the status bar.

Also, adds a compose version for the airplane mode icon.

Bug: 407813930
Test: Make sure airplane mode icon is updated
Screenshot provided in bug.
Flag: com.android.systemui.status_bar_system_status_icons_in_compose

Change-Id: I6d46eb41e09bc326c501ec9c37c07664ab40c056