Move calls to RotationPolicy to bg
When callbacks are added to RotationLockControllerImpl, it queries
RotationPolicy (Settings provider) in that thread. For RotationLockTile,
this is tied to the Lifecycle state change, so it happens in the main
thread, causing binder calls in the main thread.

Instead, move them to the background, but keep the call on the
RotationLockControllerCallback in the main thread, as consumers depend
on this.

Test: manual, toggle rotation
Test: atest RotationLockControllerImplTest
Test: perfetto, see binder calls move to background
Fixes: 407698587
Flag: EXEMPT bug fix
Change-Id: I62456f7f48253be97192e9ef11ff41b52566b8d5