Move initial value to be on main thread initially.
Move initial value to be on the main thread. This is done to avoid
holding the WindowManagerGlobal lock.
Move all array modifications to use a handler and the main thread. This
avoids holding on to the lock while modifying the list of listeners. The
list is expected to be small so it should not cause issues.

Flag: android.view.flags.root_view_changed_listener
Bug: 394397033
Test: atest CtsViewTestCases:WindowInspectorTest
Change-Id: I66a6b61afeb1f22c98963d4713e6f7543dcbd3da