Commit to the AnrTimerService
This removes the anr_timer_service flag and the code paths that rely
on it.

It is possible for the AnrTimerService to rely on MessageQueue, if the
local process does not support native timers.  This would be the case
for tests.  In that case, BroadcastQueue will not extend ANR timeouts
based on cpu load.  This should not be a problem in test scenarios.

Flag: EXEMPT flag-removal
Bug: 407063060
Test: atest
 * FrameworksServicesTests:AnrTimerTest
 * FrameworksServicesTests:com.android.server.am
 * FrameworksMockingServicesTests:com.android.server.am
Change-Id: Ie1d74c08864a101994c2b22a41a37f369949972a