Extracting display switch logic to FoldableDisplaySwitchTrackingInteractor
First step of implementing tracking unfolding velocity - we will use the same logic for tracking display switch so it's worth extracting it to separate class that can be used in both places.
For now test remains the same (to make sure everything works as before) but in the follow-up I'll split it in two separate test classes.

Bug: 405128401
Flag: EXEMPT refactor with keeping test the same
Test: DisplaySwitchLatencyTrackerTest
Change-Id: Ic1d4e7022974a6eabb9af1960fe89ed699bd5dfd