Update Perfetto file puller key pattern
The pattern is a regex expression. If any part of the metric matches the pattern we pull the file pointed to by the metric. The `*` we had in the expression just meant we would match metrics with 0+ `h` after `perfetto_file_pat`, which works but is misleading as it suggests * is a match all, which is isn't.

Bug: 404268266
Flag: TEST_ONLY
Change-Id: Ic9677b8e79d53da3826d7bbf6d20d740cf3b4989