Fix flickering of min priority notification icons
The previous logic was looking at making the dots hidden at the
state calculation level based on a state that would only be
updated later in the phase. This lead to icons flickering to
visible and hidden in a reproducible manner.
We now instead use the values during the state calculation to
hide the right icons and only have the last dot visible.

Fixes: 402752618
Test: atest SystemUITests
Flag: com.android.systemui.physical_notification_movement
Change-Id: Ife69d09a770f7fa0cd050dc5e1728d0113b13f6f