Add nullability annotations to ISecureElementService et al
The Java OMAPI implementation passes and expects null pointers for
many arguments/return values, which means the Rust native
implementation must as well to be compatible.  But the Rust AIDL
backend does not allow arguments/returns to be nullable without the
appropriate annotation.  This CL provides the required annotations.

Note that the build flag is not actually used because AIDL doesn't
support flagging.

Bug: 380331467
Test: atest CtsOmapiTestCases
Flag: build.RELEASE_NATIVE_OMAPI
Change-Id: Ia8aedeb2d1cbf6f450ae9879771a304c5be5705a