Desks: Only deactivate pre-recents desk
Ensures that the only desk that is deactivated on a
!returnToApp recents finish is the desk that was active when we first
entered Overview. This is because desk->desk transitions from Overview
are also !returnToApp, and by the time this finishing function is called
the new desk is now the "active" one, which would result in deactivating
the desk that the user intended to activate.

With this change, the active desk is saved when recents first starts, so
that only that one is deactivated on finishing if !returnToApp (e.g.
going Home or to a non-desktop tile)

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 409848289
Bug: 409639676
Test: switch desks from overview - verify activation change is
successful

Change-Id: I227b03856784796d475a9d4de003c40c1993f766