PD/CD: Always show app header if window is in freeform.
When windows are in freeform, handle is never shown - only the header.
And although in Projected Display we should hide entry points such as
the app handle to enter desktop, if the user somehow has the window in
freeform (e.g. via other means, such as adb launch), we should not hide
the header in that case, which would leave it in a rather broken UI
state.

Bug: 384568161
Test: atest
Flag:com.android.window.flags.enable_projected_desktop_mode
Change-Id: Ifa4ad9a6078d9995a1a5089d5a27c7e2f8c090a2