Fix keyguard clock flicker during Gone->Dream with shade open
In GoneToDreamingTransitionViewModel, the keyguard alpha animates out
(1f to 0f). When shade is open, keyguard clock is revealed and then
animates upwards as shade collapses to launch dream.
Fix is to set alpha to 0 if either shade or QS is expanded with an
amount on transition start.

Fixes: b/381944762
Flag: EXEMPT bugfix
Test: on mobile with shade open; tablet on dock with HUN
Test: atest GoneToDreamingTransitionViewModelTest
Change-Id: Ib587244e87aa4e9c93ce85d864ba11172bbfa633