Fix notifications showing when starting dream with shade open
If the dream starts due to the user pressing power button or the screen
timing out when the shade is open on the lock screen, the shade
collapses but the notifications remain visible until the dream has
fully started.

This CL fixes the issue, which is caused by the notifications fading
back in on collapse. This CUJ is not often used so this bug has likely
been around for a long time.

Bug: 409720192
Test: atest SharedNotificationContainerViewModelTest
Flag: NONE small animation fix
Change-Id: If7bc0dd6336fbf6d5a954636b2f1b8048215167e