Fix lack of client package name after system session updates
This fixes an issue that already existed before mirroring, but was
surfaced as a result of mirroring changes because mirroring triggers
updates in the system session.

The system session (which is populated with AudioManager routes) is
created in system_server without a package name, and then each router
takes care of populating the package name. There was a gap in this
"ensuring the client package name is populated" for system session
updates. But those are uncommon in CTS because they depend on an
AudioManager event, which doesn't happen during CTS. With mirroring,
system session updates become more frequent. This changes fixes the gap.

This gap was detected through MediaRouter2Test#testMediaRouter2ReturnsRoutingSessionInfoWithNonEmptyClientName.

Bug: b/362507305
Test: atest CtsMediaBetterTogetherTestCases:android.media.router.cts.SystemMediaRoutingTest
Test: atest CtsMediaRouterTestCases:MediaRouter2Test
Flag: com.android.media.flags.enable_mirroring_in_media_router_2
Change-Id: I7995d3ba09379f034a43dcb837db58e8efaa9942