Make a copy of surface control for bubble leash
The surface control from transition should only be used while
animating. If the component wants to keep the surface, it is
safer to make a copy and release by itself.

Also remove manual reparent to null, because the surface is the
real surface of task, core already performs the surface removal
when removing the task.

Bug: 411280316
Flag: EXEMPT bugfix
Test: Launch a bubble task.
      No crash on rotate screen and remove bubble.
Change-Id: I109735be704f4ecf61f88f43a96b4be9dfc27de0