[SB][Chips] Add OngoingActivityChipsRefiner
This CL adds a refiner feature for RON chips. An
OngoingActivityChipsRefiner takes the entire set of chips produced by
the OngoingActivityChipsViewModel and can apply a transform on them to
produce a modified set of chips.

This change itself does not result in
any change in behavior. Specific refiners that actually change chips
behavior will be added as part of new features.

Bug: 413216082
Test: OngoingActivityChipsViewModelTest
Flag: EXEMPT because this change doesn't have any impact on
chip behavior.

Change-Id: I8cb8542cdf0bdf3050c6a31b39fdcdd9f56aa65b