RecipientList&friends: virtual destructors
RecipientList inherits from RefBase, which has a virtual
destructor. So, we should also have one. Though, from
the stacktraces, it appears this is working without
one.

JavaRecipient classes also fixed in the same way.
The frozen callback is for completion. It may
matter for JavaDeathRecipient, as I haven't
seen a stack for it. Though, I also don't think this
is the issue, if that is the issue sometimes, it
should give us a better error message.

Bug: 405737267
Test: N/A
Change-Id: Ic9d9f3584253a2688bf856d357b762609b739b45