[flexi] 1f device entry icon alpha when KS.UNDEFINED
We only want to hide the device entry icon when
the Scene is Lockscreen (between transitions and on some
lockscreen keyguard transition states).

Otherwise, the device entry icon should always be 1f,
including when the KeyguardState is UNDEFINED (which means
its on a non-lockscreen scene).

Test: manually swipe down and up on the lockscreen, see
device entry icon
Fixes: 379845752
Flag: EXEMPT bugfix

Change-Id: I7767fc899fcc3cb8d8bcaf5d94c6ee250a218b6d