Ensure dispose is always called on looper thread -- try 3
The InputEventReceiver::dispose should only be called on the looper
thread. This is already documented in the class, but not enforced.

Enforce this here, since this will become critical in subsequent patches
where we are relying on the ability to consume upon dispose.

This is the third attempt, after finding a bunch of issues in the postsubmit on first and second attempts.

This reverts commit cc91b0129c8595811f70f5081ac942748b801f79.

Reason for revert: attempting again

Bug: 376713684
Test: atest InputEventSenderAndReceiverTest
Flag: EXEMPT refactor
Change-Id: Idc05f3ea5ee973d13a81ffbd422301735f6c757e