[Notif redesign] Use circular bg for left icon in messaging notifs
Since we now enforce a circular crop for avatars, we should do the same
for the left_icon since that also appears in the same fashion in
grouped messaging notifications.

Also made sure we're not actually showing the left_icon when we don't
have something to show there - since otherwise we were just displaying
an empty circle. The "apply" boolean here (remamed from showLeftIcon) is
true when we know the small icons are the same, but doesn't necessarily
mean we need to show the left icon.

Bug: 406224300
Flag: android.app.notifications_redesign_templates
Test: tested manually with a mix of conversations and non-conversations
(both individually and grouped together), in light and dark mode; also
made sure normal grouped notifs with large icons are unchanged

Change-Id: I7bae9492895976a7208a39fe1b92ed78e57e640f