Add a flag to listen to package-added in a service.
The Settings app, which restores user aspect ratio settings,
cannot set the aspect ratio for apps that are not installed.
As the settings app can be killed at any time, it is not
 suitable for listening to package-added signals.

Therefore, this will be delegated to a service, once the
ideal aspect ratio has been resolved.

Flag: EXEMPT flag creation.
Bug: 413014252
Test: N/A
Change-Id: Ieb3fb784c204cd191de7ec8983e34e4fa842f15e