Disable falsing entirely when lockscreen widgets are showing
It is difficult to open the notification shade while on the widgets on
lockscreen UI is open due to flasing. The user needs to perform a
gesture that is already checked by falsing in order to enter the page,
so just disable falsing when the hub is open.

Bug: 409682586
Test: atest FalsingCollectorImplTest BrightLineFalsingManagerTest
Flag: com.android.systemui.communal_shade_touch_handling_fixes
Change-Id: I3bc2661aeb852e2adb90547d28fa8d994087b9b0