UserController.getUserInfo outside of mLock
Moves getUserInfo() calls (which call into UserManager and obtain
mUsersLock) outside of the UserController.mLock blocks.
I don't know that doing this inside the mLock block is unsafe, but we
usually make sure to do it outside; to be safe, let's keep doing so.

Actually, getCurrentUserLU() still calls getUserInfo() and is called
inside mLock blocks, and has been that way for ages, so it does appear
to be safe. This cl doesn't touch that method.

Test: boots
Flag: EXEMPT bugfix
Change-Id: I457d1c0bb4fee2913381c24f30172984cc9b923e