Add haptics to magnetic divider
Initializes MSDLPlayer in WMShellModule as a singleton to be used in Shell. (Can be cleaned up and replaced by a global import in GlobalModule -- currently being worked on by haptics team)

Adds logic to listen to segment changes in the MotionValue controller, and plays a small haptic click when attaching or detaching from a snap target.

Bug: 383631946
Flag: com.android.wm.shell.enable_magnetic_split_divider
Test: No changes with flag off. Plays haptics with flag on.
Change-Id: If0bdc46627afdc4f1794307333f10aba96210e3f